config: return mutable object
We initially tought config data as immutable (collections.namedtuple),
which is easier to reason about, but proven to be slightly inconvenient
for the client (which lives in a mutable world), especially when only
few changes are needed.

For the sake of practicality, we now return a mutable object, and we
add safety nets in the config-handling code to make sure the new
mutability does not provide harm.